+++ modules/ccv2/core/src/sap/commerce/toolset/ccv2/dto/CCv2EnvironmentDto.kt
@@ -21,7 +21,6 @@ package sap.commerce.toolset.ccv2.dto
 import sap.commerce.toolset.ccv1.model.EnvironmentHealthDTO
 import sap.commerce.toolset.ccv2.CCv2Constants
 import sap.commerce.toolset.ccv2.model.EnvironmentDetailDTO
-import sap.commerce.toolset.ccv2.model.EnvironmentScalingDetailDTO
 import sap.commerce.toolset.ccv2.settings.state.CCv2Subscription
 
 data class CCv2EnvironmentDto(
@@ -40,8 +39,7 @@ data class CCv2EnvironmentDto(
     var services: Collection<CCv2ServiceDto>? = null,
     var dataBackups: Collection<CCv2DataBackupDto>? = null,
     var endpoints: Collection<CCv2EndpointDto>? = null,
-    var clusterScaling: Collection<CCv2ClusterScalingDto>? = null,
-    var databaseSchemas: Collection<CCv2DatabaseSchemaDto>? = null,
+    var scaling: CCv2EnvironmentScalingDto? = null,
 ) : CCv2Dto, Comparable<CCv2EnvironmentDto> {
 
     val accessible
@@ -65,7 +63,6 @@ data class CCv2EnvironmentDto(
         var canAccess: Boolean,
         var v1Environment: sap.commerce.toolset.ccv1.model.EnvironmentDTO? = null,
         var v1EnvironmentHealth: EnvironmentHealthDTO? = null,
-        var scaling: EnvironmentScalingDetailDTO? = null,
     )
 
     companion object {
@@ -98,12 +95,6 @@ data class CCv2EnvironmentDto(
                 problems = v1EnvironmentHealth?.problems,
                 link = link,
                 mediaStorages = mediaStorages,
-                clusterScaling = mappingDto.scaling
-                    ?.kubernetesClusters
-                    ?.mapNotNull { CCv2ClusterScalingDto.map(it) },
-                databaseSchemas = mappingDto.scaling
-                    ?.databaseSchemas
-                    ?.mapNotNull { CCv2DatabaseSchemaDto.map(it) },
             )
         }
     }