+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/actionSystem/CCv2DeleteBuildAction.kt
@@ -50,6 +50,11 @@ class CCv2DeleteBuildAction : CCv2Action(
         CCv2Service.getInstance(project).deleteBuild(project, subscription, build)
     }
 
-    override fun isEnabled(e: AnActionEvent) = super.isEnabled(e)
-        && (e.getData(CCv2UiConstants.DataKeys.Build)?.canDelete() ?: false)
+    override fun update(e: AnActionEvent) {
+        super.update(e)
+
+        if (e.presentation.isEnabledAndVisible) {
+            e.presentation.isVisible = e.getData(CCv2UiConstants.DataKeys.Build)?.canDelete() ?: false
+        }
+    }
 }