+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/actionSystem/CCv2FetchEnvironmentServiceAction.kt
@@ -46,7 +46,7 @@ class CCv2FetchEnvironmentServiceAction : DumbAwareAction("Fetch Service", null,
 
         fetching = true
 
-        CCv2Service.Companion.getInstance(project).fetchEnvironmentServices(
+        CCv2Service.getInstance(project).fetchEnvironmentServices(
             subscription,
             environment,
             { response ->